+++ e2e-tests/standalone-parser/src/types3.ts
@@ -18,3 +18,5 @@ export type Rate = NumberFormatExtends<NonNegativeNumber, "Rate">;
 export type UserId = StringFormat<"UserId">;
 export type ReadAuthorizedUserId = StringFormatExtends<UserId, "ReadAuthorizedUserId">;
 export type WriteAuthorizedUserId = StringFormatExtends<ReadAuthorizedUserId, "WriteAuthorizedUserId">;
+
+export type CurrencyPrices = Record<ValidCurrency, Rate>;