+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/client
 
+## 0.0.92
+
+### Patch Changes
+
+- mapped recordf
+- Updated dependencies
+  - @beff/cli@0.0.92
+
 ## 0.0.91
 
 ### Patch Changes